refactor: move myAccounts to features
