+++ src/components/new-safe/create/steps/ReviewStep/index.tsx
@@ -42,7 +42,7 @@ import { useRouter } from 'next/router'
 import { useMemo, useState } from 'react'
 import ChainIndicator from '@/components/common/ChainIndicator'
 import NetworkWarning from '../../NetworkWarning'
-import useAllSafes from '@/components/welcome/MyAccounts/useAllSafes'
+import useAllSafes from '@/features/myAccounts/hooks/useAllSafes'
 import { uniq } from 'lodash'
 import { selectRpc } from '@/store/settingsSlice'
 import { AppRoutes } from '@/config/routes'
