+++ src/features/myAccounts/components/AccountItems/AccountItem.tsx
@@ -20,8 +20,8 @@ import { sameAddress } from '@/utils/addresses'
 import classnames from 'classnames'
 import { useRouter } from 'next/router'
 import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline'
-import type { SafeItem } from './useAllSafes'
-import { useGetHref } from './useGetHref'
+import type { SafeItem } from '@/features/myAccounts/hooks/useAllSafes'
+import { useGetHref } from '@/features/myAccounts/hooks/useGetHref'
 import { extractCounterfactualSafeSetup, isPredictedSafeProps } from '@/features/counterfactual/utils'
 import useWallet from '@/hooks/wallets/useWallet'
 import { hasMultiChainAddNetworkFeature } from '@/features/multichain/utils/utils'
@@ -33,7 +33,7 @@ import useOnceVisible from '@/hooks/useOnceVisible'
 import { skipToken } from '@reduxjs/toolkit/query'
 import { defaultSafeInfo, useGetSafeOverviewQuery } from '@/store/slices'
 import FiatValue from '@/components/common/FiatValue'
-import QueueActions from './QueueActions'
+import QueueActions from '../QueueActions'
 import VisibilityIcon from '@mui/icons-material/Visibility'
 
 type AccountItemProps = {
