+++ src/features/myAccounts/components/AccountItems/SubAccountItem.tsx
@@ -19,10 +19,10 @@ import { sameAddress } from '@/utils/addresses'
 import classnames from 'classnames'
 import { useRouter } from 'next/router'
 import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline'
-import type { SafeItem } from './useAllSafes'
+import type { SafeItem } from '@/features/myAccounts/hooks/useAllSafes'
 import FiatValue from '@/components/common/FiatValue'
-import QueueActions from './QueueActions'
-import { useGetHref } from './useGetHref'
+import QueueActions from '@/features/myAccounts/components/QueueActions'
+import { useGetHref } from '@/features/myAccounts/hooks/useGetHref'
 import { extractCounterfactualSafeSetup } from '@/features/counterfactual/utils'
 import VisibilityIcon from '@mui/icons-material/Visibility'
 import useOnceVisible from '@/hooks/useOnceVisible'
