+++ src/features/myAccounts/components/SafesList/index.tsx
@@ -1,7 +1,7 @@
-import AccountItem from './AccountItem'
-import { type SafeItem } from './useAllSafes'
-import { type MultiChainSafeItem } from './useAllSafesGrouped'
-import MultiAccountItem from './MultiAccountItem'
+import AccountItem from '@/features/myAccounts/components/AccountItems/AccountItem'
+import type { SafeItem } from '@/features/myAccounts/hooks/useAllSafes'
+import type { MultiChainSafeItem } from '@/features/myAccounts/hooks/useAllSafesGrouped'
+import MultiAccountItem from '@/features/myAccounts/components/AccountItems/MultiAccountItem'
 import { isMultiChainSafeItem } from '@/features/multichain/utils/utils'
 import { TransitionGroup } from 'react-transition-group'
 import { Collapse } from '@mui/material'
